put connection establishment structs in coreclient
